+++ app/Http/Controllers/UserController.php
@@ -1,68 +1,86 @@
 <?php
 
 namespace App\Http\Controllers;
-use Illuminate\Support\Facades\Hash;
+
 use App\Http\Requests\StoreUpdateUserFormRequest;
 use App\Models\User;
-use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 
 class UserController extends Controller
 {
+    protected $model;
 
-    public function index() :JsonResponse
+    public function __construct(User $user)
     {
-
-        return response()->json([
-            'usuarios' => User::all(),
-        ], 200);
+        $this->model = $user;
     }
 
-    public function show(User $id) :JsonResponse
+    public function index(Request $request)
     {
-        return response()->json($id, 200);
+        $users = $this->model
+                        ->getUsers(
+                            search: $request->search ?? ''
+                        );
+
+
+        return view('users.index', compact('users'));
+    }
+
+    public function show($id)
+    {
+        // $user = $this->model->where('id', '=', $id)->first();
+        if (!$user = $this->model->find($id))
+            return redirect()->route('users.index');
+
+        return view('users.show', compact('user'));
     }
 
     public function create()
     {
-        //create = store
+        return view('users.create');
     }
 
-    public function store(Request $request) :JsonResponse
+    public function store(StoreUpdateUserFormRequest $request)
     {
-        $u = (object)$request->all();
-        
-        $usuario = new User;
-        $usuario->name = $u->name;
-        $usuario->email = $u->email;
-        $usuario->password = Hash::make($u->password);
-        $usuario->save();
-        return response ()->json($usuario, 200);
+        $data = $request->all();
+        $data['password'] = bcrypt($request->password);
+
+        $user = new User();
+        return $user->storeUser($data);
+
     }
 
-    // public function edit(User $user) :JsonResponse
-    // {    
-    //     return response()->json($user, 200);
-    // }
-
-    public function update(Request $request, User $user) :JsonResponse
+    public function edit($id)
     {
-        $data = [
-            'name' => $request->name,
-            'email' => $request->email,
-            'password' => $request->password
-        ];
+        if (!$user = $this->model->find($id))
+            return redirect()->route('users.index');
+
+        return view('users.edit', compact('user'));
+    }
+
+    public function update(StoreUpdateUserFormRequest $request, $id)
+    {
+        if (!$user = $this->model->find($id))
+            return redirect()->route('users.index');
+
+        $data = $request->only('name', 'email');
+        if($request->password)
+            $data['password'] = bcrypt($request->password);
+
         $user->update($data);
-        return response ()->json($user, 200);
+
+        return redirect()->route('users.index');
     }
 
-    public function delete(User $user) :JsonResponse
+    public function delete($id)
     {
+        if (!$user = $this->model->find($id))
+            return redirect()->route('users.index');
+
         $user->delete();
-        return response ()->json('Registro deletado', 200);
+
+        return redirect()->route('users.index');
     }
-
-
 
 
 }
